+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Changelog
 ---------
 
 v1.31 (??? 2023)
+    * Allow ncol=1 in `color.get_cmap`.
+    * prepend and append keywords in `mad`.
     * Warn only if lb > ub, simply set mask if lb == ub in `sce`.
     * Add functions `pack`, `unpack` mimicking Fortran intrinsics.